+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+* Omnia: Improve labels of USB ports (front/rear).
+
 ## [0.11.0] - 2023-02-22
 ### Added
 * Add optional info about pci slot path for wireless devices.