+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * utils: Match the QMI control device to its interface and report it in
   resulting json data.
 * tests: Add mock data for Mox ABC with two QMI modems.
+* tests: Add mock data for Turris 1.1 with two QMI modems.
 
 ### Changed
 * Omnia: Improve labels of USB ports (front/rear).